refactor(RoomView): gate blocking screens before mounting the room tree
RoomView's blocked-room checks ran after twelve hooks, so an invited or E2EE-blocked room opened the DDP subscription and ran init before being told it was blocked, and unblocking swapped component types at the same render position.

index.tsx is now a thin RoomGate that owns the screen-identity snapshot, the room store acquisition, the header and the blocked-screen checks; the room tree moves to RoomScreen.tsx and mounts only once unblocked. RoomLoadFailed stays in RoomScreen: it needs useRoomInit's failed/retry, which belongs to the mounted room.
